HW-02

+++ src/main/java/ru/javawebinar/topjava/repository/mock/InMemoryMealRepositoryImpl.java
@@ -1,20 +1,26 @@
 package ru.javawebinar.topjava.repository.mock;
 
+import org.springframework.stereotype.Repository;
+import ru.javawebinar.topjava.AuthorizedUser;
 import ru.javawebinar.topjava.model.Meal;
 import ru.javawebinar.topjava.repository.MealRepository;
+import ru.javawebinar.topjava.util.DateTimeUtil;
 import ru.javawebinar.topjava.util.MealsUtil;
 
-import java.util.Collection;
-import java.util.Map;
+import java.time.LocalDate;
+import java.util.*;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.stream.Collectors;
 
+@Repository
 public class InMemoryMealRepositoryImpl implements MealRepository {
     private Map<Integer, Meal> repository = new ConcurrentHashMap<>();
     private AtomicInteger counter = new AtomicInteger(0);
 
     {
         MealsUtil.MEALS.forEach(this::save);
+        repository.get(1).setUserId(2);
     }
 
     @Override
@@ -22,23 +28,35 @@ public class InMemoryMealRepositoryImpl implements MealRepository {
         if (meal.isNew()) {
             meal.setId(counter.incrementAndGet());
         }
+        Meal mealInRep=repository.get(meal.getId());
+        if (mealInRep!=null&&mealInRep.getUserId()!=AuthorizedUser.id()) return null;
         repository.put(meal.getId(), meal);
+        meal.setUserId(AuthorizedUser.id());
         return meal;
     }
 
     @Override
-    public void delete(int id) {
+    public boolean delete(int id) {
+        Meal meal = repository.get(id);
+        if (meal == null || meal.getUserId() != AuthorizedUser.id()) return false;
         repository.remove(id);
+        return true;
     }
 
     @Override
     public Meal get(int id) {
-        return repository.get(id);
+        Meal meal = repository.get(id);
+        if (meal == null || meal.getUserId() != AuthorizedUser.id()) return null;
+        return meal;
     }
 
     @Override
-    public Collection<Meal> getAll() {
-        return repository.values();
+    public Collection<Meal> getAll(LocalDate start, LocalDate end) {
+        return repository.values().stream()
+                .filter(m -> m.getUserId() == AuthorizedUser.id())
+                .filter(m-> (DateTimeUtil.isBetween(m.getDate(),start,end)))
+                .sorted(Comparator.comparing(Meal::getDateTime).reversed())
+                .collect(Collectors.toList());
     }
 }
